pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-fm"
version = "0.1.0"
description = "CLI tool for managing Obsidian notes frontmatter"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "User"}
]
dependencies = [
    "click>=8.1.0",
    "python-frontmatter>=1.0.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
obsidian-fm = "obsidian_fm.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
